config: Update Jenkins config
This commit updates the Jenkins config after we have introduced
Fedora 26 as a new distribution being tested through our CI.

It also updates all the job names so that we can make the
difference between a pull request and a merge to master.

And it does now include the users configurations.

Fixes #10

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>