jobs: Update Jenkins config to limit logs
Those jobs have been updated because we wanted to store the extra
logs as build artifacts instead of making them part of the console
output. That way, we avoid web browser crashing when trying to display
the logs, and it is easier to find where the failure comes from.

This commit also updates the plugins.

Fixes #20

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>